+++ src/Controller/Api/PlaylistSlideDeleteController.php
@@ -7,6 +7,7 @@ namespace App\Controller\Api;
 use App\Repository\PlaylistSlideRepository;
 use App\Utils\ValidationUtils;
 use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Attribute\AsController;
 
 #[AsController]
@@ -24,6 +25,6 @@ class PlaylistSlideDeleteController extends AbstractTenantAwareController
 
         $this->playlistSlideRepository->deleteRelations($ulid, $slideUlid, $this->getActiveTenant());
 
-        return new JsonResponse(null, \Symfony\Component\HttpFoundation\Response::HTTP_NO_CONTENT);
+        return new JsonResponse(null, Response::HTTP_NO_CONTENT);
     }
 }