+++ src/Repository/ThemeRepository.php
@@ -17,7 +17,7 @@ use Symfony\Component\Uid\Ulid;
  * @method Theme[]    findAll()
  * @method Theme[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
  *
- * @extends \Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository<\App\Entity\Tenant\Theme>
+ * @extends ServiceEntityRepository<Theme>
  */
 class ThemeRepository extends ServiceEntityRepository
 {